+++ PKHeX.Core/Legality/Ribbons/RibbonStrings.cs
